refactor: Update component imports and file structure
This commit refactors the component imports and file structure to improve organization and maintainability. The changes include:
- Moving the `AboutSection` component from `src/components/AboutSection.vue` to `src/components/sections/AboutSection.vue`
- Updating the import paths in `Home.vue` to reflect the new file structure
- Removing the `HeaderNav` component from `AppHeader/HeaderNav.vue`

These changes were made to align with established conventions and improve the overall structure of the codebase.
